Squashed commit of the following:
commit 2cd847b5ec602973376c0fd13437689e044aaf57
Merge: f60be42b97 9020abff2b
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Dec 17 22:22:12 2024 -0500

    Merge branch 'main' into jongsun/perf/redux/241122-move-metamask-slice-ui-properties-to-appstate-slice

commit f60be42b976ffdac2e39584b53abe1c249d46fab
Merge: fd07d5a7f0 25d1b62fa3
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Wed Dec 11 18:02:08 2024 -0500

    Merge branch 'main' into jongsun/perf/redux/241122-move-metamask-slice-ui-properties-to-appstate-slice

commit fd07d5a7f0ad4a2cfa8cc71c2bb787b14614c750
Merge: f2da87caa4 07d4265d16
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Wed Dec 11 09:27:20 2024 -0500

    Merge branch 'main' into jongsun/perf/redux/241122-move-metamask-slice-ui-properties-to-appstate-slice

commit f2da87caa42d32f5bd39f1a53b7bff3dccfb2b9d
Merge: 537fef8564 f9e7878966
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Dec 10 09:34:11 2024 -0500

    Merge branch 'main' into jongsun/perf/redux/241122-move-metamask-slice-ui-properties-to-appstate-slice

commit 537fef85642892e9ae1bf36fd8a0f811ffe6a428
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Fri Dec 6 13:03:15 2024 -0500

    Group `appState` selectors in `ui/selectors/selectors/js` file

commit 670be0aece2cdb56cadc66c309bdcabeba749b53
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Fri Dec 6 11:42:53 2024 -0500

    Clean up traces of removed `appState` property `gasIsLoading`

commit ad7eed860ecc6d35780972ee7f256fbebae29685
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Dec 2 13:50:47 2024 -0500

    e2e: fix `errors-after-init-opt-in-ui-state`

commit ca240215735f91dff475ea759ed99edde3f85127
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Dec 2 12:02:20 2024 -0500

    Update test objects

commit 773f3457ecfe173ffc3edbdb5eb92b13c18e2ce3
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Dec 2 05:52:34 2024 -0500

    Update sentry UI_STATE object

commit 6f908cf6552450c4f84d8bc591dd6aaa520324a5
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Dec 2 05:52:18 2024 -0500

    Update `mapStateToProps`

commit 2cc116e1e18962c9ec878630aa9ffd4e4bed96df
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Dec 2 05:10:46 2024 -0500

    Move `isInitialized` back to `metamask` slice

commit 8899b38a888ee8b16a4d9c9019c1f7710e2244d9
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Nov 26 05:42:20 2024 -0500

    Fix assets test

commit 071a7b6a3e2ba14cb348436e084ebae4855315df
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Nov 26 05:42:11 2024 -0500

    Fix confirmations test

commit 236191afa2c12373615ad31edc00171b273ef098
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Nov 26 05:07:13 2024 -0500

    Narrow `unknown` entries in `AppState` type

commit 27acc98427b42aea9755e3e8c4bd9b71c5d5ad43
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Tue Nov 26 04:56:12 2024 -0500

    fix breaking tests

commit 3e9ed7c4d2f8367481ae6e7468cb1225a3265029
Author: Jongsun Suh <jongsun.suh@icloud.com>
Date:   Mon Nov 25 15:19:16 2024 -0500

    Move non controller-derived properties from `metamask` slice to `appState` slice
