+++ test/e2e/tests/metrics/state-snapshots/errors-after-init-opt-in-ui-state.json
@@ -1,7 +1,71 @@
 {
   "DNS": "object",
   "activeTab": "object",
-  "appState": "object",
+  "appState": {
+    "customNonceValue": "",
+    "isAccountMenuOpen": false,
+    "isNetworkMenuOpen": false,
+    "nextNonce": null,
+    "confirmationExchangeRates": {},
+    "ledgerTransportStatus": "string",
+    "ledgerWebHidConnectedStatus": "string",
+    "loadingMessage": "undefined",
+    "menuOpen": "boolean",
+    "modal": "object",
+    "alertOpen": "boolean",
+    "alertMessage": null,
+    "networkDropdownOpen": "boolean",
+    "importNftsModal": "object",
+    "keyringRemovalSnapModal": "object",
+    "importTokensModalOpen": "boolean",
+    "deprecatedNetworkModalOpen": "boolean",
+    "accountDetail": "object",
+    "isLoading": "boolean",
+    "isNftStillFetchingIndication": "boolean",
+    "buyView": "object",
+    "defaultHdPaths": "object",
+    "networksTabSelectedRpcUrl": "string",
+    "requestAccountTabs": "object",
+    "currentWindowTab": "object",
+    "gasLoadingAnimationIsShowing": "boolean",
+    "externalServicesOnboardingToggleState": "boolean",
+    "newNftAddedMessage": "string",
+    "removeNftMessage": "string",
+    "newNetworkAddedName": "string",
+    "editedNetwork": "undefined",
+    "newNetworkAddedConfigurationId": "string",
+    "newTokensImported": "string",
+    "newTokensImportedError": "string",
+    "onboardedInThisUISession": "boolean",
+    "customTokenAmount": "string",
+    "accountDetailsAddress": "string",
+    "isAddingNewNetwork": "boolean",
+    "isMultiRpcOnboarding": "boolean",
+    "errorInSettings": null,
+    "openMetaMaskTabs": "object",
+    "pendingTokens": "object",
+    "qrCodeData": null,
+    "selectedNetworkConfigurationId": "string",
+    "scrollToBottom": "boolean",
+    "sendInputCurrencySwitched": "boolean",
+    "shouldClose": "boolean",
+    "showBasicFunctionalityModal": "boolean",
+    "showDataDeletionErrorModal": "boolean",
+    "showDeleteMetaMetricsDataModal": "boolean",
+    "showIpfsModalOpen": "boolean",
+    "showKeyringRemovalSnapModal": "boolean",
+    "showNftDetectionEnablementToast": "boolean",
+    "showPermittedNetworkToastOpen": "boolean",
+    "showTermsOfUsePopup": "boolean",
+    "showWhatsNewPopup": "boolean",
+    "singleExceptions": "object",
+    "smartTransactionsError": null,
+    "smartTransactionsErrorMessageDismissed": "boolean",
+    "snapsInstallPrivacyWarningShown": "boolean",
+    "txId": null,
+    "warning": "string",
+    "welcomeScreenSeen": false
+  },
   "bridge": "object",
   "confirmAlerts": "object",
   "confirmTransaction": "object",
@@ -12,18 +76,12 @@
   "metamask": {
     "isInitialized": true,
     "isUnlocked": true,
-    "isAccountMenuOpen": false,
-    "isNetworkMenuOpen": false,
     "internalAccounts": { "accounts": "object", "selectedAccount": "string" },
     "transactions": "object",
     "networkConfigurations": "object",
     "addressBook": "object",
-    "confirmationExchangeRates": {},
-    "pendingTokens": "object",
-    "customNonceValue": "",
     "useBlockie": false,
     "featureFlags": {},
-    "welcomeScreenSeen": false,
     "currentLocale": "en",
     "preferences": {
       "hideZeroBalanceTokens": false,
@@ -49,7 +107,6 @@
     "use4ByteResolution": true,
     "participateInMetaMetrics": true,
     "dataCollectionForMarketing": "boolean",
-    "nextNonce": null,
     "currencyRates": {
       "ETH": {
         "conversionDate": "number",
