+++ ui/ducks/metamask/metamask.js
@@ -28,18 +28,12 @@ import { setCustomGasLimit, setCustomGasPrice } from '../gas/gas.duck';
 const initialState = {
   isInitialized: false,
   isUnlocked: false,
-  isAccountMenuOpen: false,
-  isNetworkMenuOpen: false,
   internalAccounts: { accounts: {}, selectedAccount: '' },
   transactions: [],
   networkConfigurations: {},
   addressBook: [],
-  confirmationExchangeRates: {},
-  pendingTokens: {},
-  customNonceValue: '',
   useBlockie: false,
   featureFlags: {},
-  welcomeScreenSeen: false,
   currentLocale: '',
   currentBlockGasLimit: '',
   currentBlockGasLimitByChainId: {},
@@ -60,7 +54,6 @@ const initialState = {
   use4ByteResolution: true,
   participateInMetaMetrics: null,
   dataCollectionForMarketing: null,
-  nextNonce: null,
   currencyRates: {
     ETH: {
       conversionRate: null,
@@ -121,24 +114,6 @@ export default function reduceMetamask(state = initialState, action) {
       return Object.assign(metamaskState, { internalAccounts });
     }
 
-    case actionConstants.UPDATE_CUSTOM_NONCE:
-      return {
-        ...metamaskState,
-        customNonceValue: action.value,
-      };
-
-    case actionConstants.TOGGLE_ACCOUNT_MENU:
-      return {
-        ...metamaskState,
-        isAccountMenuOpen: !metamaskState.isAccountMenuOpen,
-      };
-
-    case actionConstants.TOGGLE_NETWORK_MENU:
-      return {
-        ...metamaskState,
-        isNetworkMenuOpen: !metamaskState.isNetworkMenuOpen,
-      };
-
     case actionConstants.UPDATE_TRANSACTION_PARAMS: {
       const { id: txId, value } = action;
       let { transactions } = metamaskState;
@@ -169,25 +144,6 @@ export default function reduceMetamask(state = initialState, action) {
         dataCollectionForMarketing: action.value,
       };
 
-    case actionConstants.CLOSE_WELCOME_SCREEN:
-      return {
-        ...metamaskState,
-        welcomeScreenSeen: true,
-      };
-
-    case actionConstants.SET_PENDING_TOKENS:
-      return {
-        ...metamaskState,
-        pendingTokens: { ...action.payload },
-      };
-
-    case actionConstants.CLEAR_PENDING_TOKENS: {
-      return {
-        ...metamaskState,
-        pendingTokens: {},
-      };
-    }
-
     case actionConstants.COMPLETE_ONBOARDING: {
       return {
         ...metamaskState,
@@ -198,13 +154,12 @@ export default function reduceMetamask(state = initialState, action) {
     case actionConstants.RESET_ONBOARDING: {
       return {
         ...metamaskState,
+        isInitialized: false,
         completedOnboarding: false,
         firstTimeFlowType: null,
-        isInitialized: false,
         isUnlocked: false,
         onboardingTabs: {},
         seedPhraseBackedUp: null,
-        welcomeScreenSeen: false,
       };
     }
 
@@ -214,18 +169,6 @@ export default function reduceMetamask(state = initialState, action) {
         firstTimeFlowType: action.value,
       };
     }
-
-    case actionConstants.SET_NEXT_NONCE: {
-      return {
-        ...metamaskState,
-        nextNonce: action.payload,
-      };
-    }
-    case actionConstants.SET_CONFIRMATION_EXCHANGE_RATES:
-      return {
-        ...metamaskState,
-        confirmationExchangeRates: action.value,
-      };
 
     default:
       return metamaskState;
@@ -283,8 +226,6 @@ export const getWeb3ShimUsageAlertEnabledness = (state) =>
 
 export const getUnconnectedAccountAlertShown = (state) =>
   state.metamask.unconnectedAccountAlertShownOrigins;
-
-export const getPendingTokens = (state) => state.metamask.pendingTokens;
 
 export const getTokens = (state) => state.metamask.tokens;
