+++ ui/pages/asset/components/asset-page.test.tsx
@@ -55,6 +55,9 @@ describe('AssetPage', () => {
     localeMessages: {
       currentLocale: 'en',
     },
+    appState: {
+      confirmationExchangeRates: {},
+    },
     metamask: {
       tokenList: {},
       tokenBalances: {
