+++ ui/pages/routes/routes.container.js
@@ -102,8 +102,8 @@ function mapStateToProps(state) {
     forgottenPassword: state.metamask.forgottenPassword,
     isCurrentProviderCustom: isCurrentProviderCustom(state),
     completedOnboarding,
-    isAccountMenuOpen: state.metamask.isAccountMenuOpen,
-    isNetworkMenuOpen: state.metamask.isNetworkMenuOpen,
+    isAccountMenuOpen: state.appState.isAccountMenuOpen,
+    isNetworkMenuOpen: state.appState.isNetworkMenuOpen,
     isImportTokensModalOpen: state.appState.importTokensModalOpen,
     isBasicConfigurationModalOpen: state.appState.showBasicFunctionalityModal,
     isDeprecatedNetworkModalOpen: state.appState.deprecatedNetworkModalOpen,
